+++ backend/src/main/java/com/juliy/store/backend/security/config/SecurityConfig.java
@@ -47,6 +47,7 @@ public class SecurityConfig {
                         .requestMatchers(HttpMethod.POST, "/api/products/**").hasRole("ADMIN")
                         .requestMatchers(HttpMethod.PUT, "/api/products/**").hasRole("ADMIN")
                         .requestMatchers(HttpMethod.DELETE, "/api/products/**").hasRole("ADMIN")
+                        .requestMatchers(HttpMethod.GET, "/api/metrics/**").hasAnyRole("ADMIN", "CAJERO")
                         .requestMatchers(HttpMethod.GET, "/api/sales/**").hasAnyRole("ADMIN", "CAJERO")
                         .requestMatchers(HttpMethod.POST, "/api/sales/**").hasAnyRole("ADMIN", "CAJERO")
                         .anyRequest().authenticated()